refactor: remove element-style on heat-map-cell (#1653)
* refactor: remove element-style on heat-map-cell:

- this allows to target `.cell` without using `!important`
- it follow the pattern which is also used for e.g. `.bar`

* refactor: remove element-style on tree-map-cell
