+++ projects/swimlane/ngx-charts/src/lib/common/base-chart.component.scss
@@ -3,6 +3,7 @@
   overflow: visible;
 
   .circle,
+  .cell,
   .bar,
   .arc {
     cursor: pointer;
